requirements.txt
Pyttsx3
PyPDF2
PyPDF2
pillow
tk
